feat: adiciona metodos buscar, atualizar e deletar
