+++ src/models/Categoria.ts
