--- src/services/Service.tsx
@@ -1,15 +1,0 @@
-import axios from 'axios'
-
-const api = axios.create({
-    baseURL: 'https://centro-bell-hooks.onrender.com/',
-})
-
-export const auth = async (url: string, dados: object, setDados: (dados: any) => void) => {
-    const resposta = await api.post(url, dados)
-    setDados(resposta.data)
-}
-
-export const cadastrarCategoria = async (url: string, dados: object, setDados: (dados: any) => void, header: object) => {
-    const resposta = await api.post(url, dados, header);
-    setDados(resposta.data);
-}
